Moved auth_toggles to companies.py, updated unit tests. removed direct import of requests and HTTPResponse
