+++ tests/test_gremlinapi.py
@@ -194,19 +194,3 @@ class TestAPI(unittest.TestCase):
             GremlinAPI._warn_if_not_param("fakeparam", test_param, **mock_identifier),
             test_param,
         )
-
-    @patch("requests.post")
-    def test_toggles_with_decorator(self, mock_get) -> None:
-        toggles_body = {
-            "companyId": hooli_id,
-            "passwordEnabled": True,
-            "mfaRequired": False,
-            "googleEnabled": True,
-            "oauthEnabled": True,
-            "samlEnabled": True,
-            "claimsRequired": False,
-        }
-        mock_get.return_value = requests.Response()
-        mock_get.return_value.status_code = 200
-        mock_get.return_value.json = mock_json
-        self.assertEqual(GremlinAPI.auth_toggles(**toggles_body), mock_get.return_value)
